% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.anesthesia-package.R
\docType{package}
\name{rc.anesthesia-package}
\alias{rc.anesthesia}
\alias{rc.anesthesia-package}
\title{rc.anesthesia: Research Compendium of the GIPSO Anesthesia Project}
\description{
Research compendium of the Interdisciplinary Sleep Research Group (GIPSO) anesthesia project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/rc.anesthesia/}
  \item \url{https://github.com/gipso/rc.anesthesia/}
  \item Report bugs at \url{https://github.com/gipso/rc.anesthesia/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID})

Authors:
\itemize{
  \item Alisson da Silva Rodrigues dos Santos \email{alisson.rodriguesds@gmail.com} (\href{https://orcid.org/0000-0002-5455-3562}{ORCID}) [conceptor, copyright holder]
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
